Assignment-02/requirements.txt
requests==2.31.0
argparse==1.4.0
py-dotenv==0.1
fastapi==0.110.0
